+++ database/tables/jomres_rates.php
@@ -4,7 +4,7 @@
  *
  * @author Vince Wooll <sales@jomres.net>
  *
- *  @version Jomres 10.4.0 (Platty Joobs edition)
+ *  @version Jomres 10.5.0
  *
  * @copyright	2005-2022 Vince Wooll
  * Jomres (tm) PHP, CSS & Javascript files are released under both MIT and GPL2 licenses. This means that you can choose the license that best suits your project, and use it accordingly
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS `#__jomres_rates` (
 	`rate_description` TEXT,
 	`validfrom` CHAR(10),
 	`validto` CHAR(10),
-	`roomrateperday` DOUBLE(11,2) NOT NULL DEFAULT 0,
+	`roomrateperday` DOUBLE(11,7) NOT NULL DEFAULT 0,
 	`extra_guests_price` DOUBLE(11,2) NOT NULL DEFAULT 0,
 	`modifiers` TEXT,
 	`mindays` SMALLINT(5),
